Add support for unsigned 16 and 32 bit field values beyond Short.MAX_VALUE and Integer.MAX_VALUE

+++ test/lib/jdk/test/lib/zink/ExtField.java
@@ -33,7 +33,7 @@ public sealed interface ExtField permits ExtZip64, ExtTs, ExtWinNT, ExtGeneric {
         return fields;
     }
 
-    short id();
-    short dsize();
+    int id();
+    int dsize();
     byte[] data();
 }
